Replaced NexaCtrl with NexaTransmitter, which works for on/off/dimming. Not enabled yet
NexaTransmitter lib from https://github.com/erix/NexaTransmitter works on the Pro Micro
without running into memory usage concerns, because NexaTransmitter doesn't store
the full command in an array before executing it - it calculates what it needs to transmit
on the fly. The interface is a bit rough around the edges at the moment, as there's only
one setSwitch method which does dimming and power on/off. It also has no integration
for flashing an LED when transmitting, nor changing the transmitter ID after instantiating
the object.

At present the library is included, but there's no way to use it (it still just controls
Maplin switches). Integrating it will require updating the state machine to accept the
commands, but that's about it.